fix tests config; code reorganization
