+++ src/main/java/com/snowflake/kafka/connector/internal/streaming/SnowflakeSinkServiceV2.java
@@ -6,6 +6,7 @@ import static com.snowflake.kafka.connector.internal.streaming.StreamingUtils.ST
 
 import com.codahale.metrics.MetricRegistry;
 import com.google.common.annotations.VisibleForTesting;
+import com.snowflake.kafka.connector.SchematizationUtils;
 import com.snowflake.kafka.connector.SnowflakeSinkConnectorConfig;
 import com.snowflake.kafka.connector.Utils;
 import com.snowflake.kafka.connector.dlq.KafkaRecordErrorReporter;
@@ -510,11 +511,11 @@ public class SnowflakeSinkServiceV2 implements SnowflakeSinkService {
     } else {
       LOGGER.info("Creating new table {}.", tableName);
       if (this.enableSchematization) {
-        if (Utils.usesAvroValueConverter(connectorConfig)) {
+        if (SchematizationUtils.usesAvroValueConverter(connectorConfig)) {
           // if we could read schema from schema registry
           this.conn.createTableWithSchema(
               tableName,
-              Utils.getSchemaMapForTable(
+              SchematizationUtils.getSchemaMapForTable(
                   tableName,
                   this.topicToTableMap,
                   this.connectorConfig.get(
