+++ src/test/java/com/snowflake/kafka/connector/UtilsTest.java
@@ -1,5 +1,6 @@
 package com.snowflake.kafka.connector;
 
+import com.snowflake.kafka.connector.internal.SchematizationTestUtils;
 import com.snowflake.kafka.connector.internal.SnowflakeErrors;
 import com.snowflake.kafka.connector.internal.TestUtils;
 import io.confluent.kafka.schemaregistry.avro.AvroSchema;
@@ -141,14 +142,16 @@ public class UtilsTest {
   public void testCollectSchemaFromTopics() throws Exception {
     SchemaRegistryClient schemaRegistry = new MockSchemaRegistryClient();
     schemaRegistry.register(
-        "topic0-value", new AvroSchema(TestUtils.AVRO_SCHEMA_FOR_SCHEMA_COLLECTION_0));
+        "topic0-value",
+        new AvroSchema(SchematizationTestUtils.AVRO_SCHEMA_FOR_SCHEMA_COLLECTION_0));
     schemaRegistry.register(
-        "topic1-value", new AvroSchema(TestUtils.AVRO_SCHEMA_FOR_SCHEMA_COLLECTION_1));
+        "topic1-value",
+        new AvroSchema(SchematizationTestUtils.AVRO_SCHEMA_FOR_SCHEMA_COLLECTION_1));
     Map<String, String> topicToTableMap = new HashMap<>();
     topicToTableMap.put("topic0", "table");
     topicToTableMap.put("topic1", "table");
     Map<String, String> schemaMap =
-        Utils.getSchemaMapForTableWithSchemaRegistryClient(
+        SchematizationUtils.getSchemaMapForTableWithSchemaRegistryClient(
             "table", topicToTableMap, schemaRegistry);
 
     assert schemaMap.get("ID").equals("int");
