+++ test/rest_request_template/travis_correct_auto_table_creation.json
@@ -2,7 +2,6 @@
   "name": "SNOWFLAKE_CONNECTOR_NAME",
   "config": {
     "connector.class": "com.snowflake.kafka.connector.SnowflakeSinkConnector",
-    "snowflake.role.name": "TESTROLE_KAFKA",
     "topics": "SNOWFLAKE_TEST_TOPIC",
     "tasks.max": "1",
     "buffer.flush.time": "10",
