+++ test/rest_request_template/travis_correct_multiple_topic_to_one_table_snowpipe.json
@@ -2,7 +2,6 @@
   "name": "SNOWFLAKE_CONNECTOR_NAME",
   "config": {
     "connector.class": "com.snowflake.kafka.connector.SnowflakeSinkConnector",
-    "snowflake.role.name": "TESTROLE_KAFKA",
     "topics": "SNOWFLAKE_TEST_TOPIC0,SNOWFLAKE_TEST_TOPIC1,SNOWFLAKE_TEST_TOPIC2",
     "snowflake.topic2table.map": "SNOWFLAKE_TEST_TOPIC0:SNOWFLAKE_CONNECTOR_NAME,SNOWFLAKE_TEST_TOPIC1:SNOWFLAKE_CONNECTOR_NAME,SNOWFLAKE_TEST_TOPIC2:SNOWFLAKE_CONNECTOR_NAME",
     "tasks.max": "1",
